+++ process/models/pfcoil.py
@@ -3461,7 +3461,7 @@ class CSCoil(Model):
                 r_cs_outer=self.data.pf_coil.r_cs_outer,
                 j_cs=self.data.pf_coil.j_cs_pulse_start,
                 b_cs_inner=self.data.pf_coil.b_cs_peak_pulse_start,
-                f_poisson_cs_structure=tfv.poisson_steel,
+                f_poisson_cs_structure=self.data.tfcoil.poisson_steel,
             )
 
             # In reality this is practially 0
@@ -3471,7 +3471,7 @@ class CSCoil(Model):
                 r_cs_outer=self.data.pf_coil.r_cs_outer,
                 j_cs=self.data.pf_coil.j_cs_pulse_start,
                 b_cs_inner=self.data.pf_coil.b_cs_peak_pulse_start,
-                f_poisson_cs_structure=tfv.poisson_steel,
+                f_poisson_cs_structure=self.data.tfcoil.poisson_steel,
             )
 
             # Allowable (hoop) stress (Pa) stress_cs_steel_max
@@ -4125,8 +4125,8 @@ class CSCoil(Model):
 
         return s_axial, forc_z_cs_self_peak_midplane
 
+    @staticmethod
     def calculate_cs_self_axial_stress(
-        self,
         z_stress_point: float,
         r_cs_outer: float,
         dz_cs_half: float,
@@ -4396,581 +4396,6 @@ class CSCoil(Model):
         hp_term_4 = alpha**2 + 1.0e0 - alpha**2 / epsilon**2 - epsilon**2
 
         return hp_term_1 * hp_term_2 - hp_term_3 * hp_term_4
-
-    def plot_cs_hoop_stress_profile(
-        self,
-        axis: plt.Axes,
-        mfile: MFile,
-        scan: int,
-        j_cs: float,
-        b_cs_inner: float,
-    ):
-        r_cs_inner = mfile.get("r_cs_inner", scan=scan)
-        r_cs_outer = mfile.get("r_cs_outer", scan=scan)
-
-        radii = np.linspace(r_cs_inner, r_cs_outer, num=10)
-        stress_values = np.array([
-            self.calculate_cs_hoop_stress(
-                r_stress_point=radius,
-                r_cs_inner=r_cs_inner,
-                r_cs_outer=r_cs_outer,
-                j_cs=j_cs,
-                b_cs_inner=b_cs_inner,
-                f_poisson_cs_structure=tfv.poisson_steel,
-                f_a_cs_turn_steel=mfile.get("f_a_cs_turn_steel", scan=scan),
-            )
-            for radius in radii
-        ])
-
-        axis.plot(
-            radii,
-            stress_values / 1e6,
-            linewidth=2,
-            label="$\\sigma_{\\theta}$,Hoop Stress",
-        )
-        max_idx = np.argmax(np.abs(stress_values))
-        max_radius = radii[max_idx]
-        max_stress = stress_values[max_idx] / 1e6
-        axis.axvline(max_radius, color="black", linestyle="--", linewidth=1.0, alpha=0.7)
-        axis.axhline(max_stress, color="black", linestyle="--", linewidth=1.0, alpha=0.7)
-        axis.set_xlabel("Radial Position (m)")
-        axis.set_ylabel("Hoop Stress (MPa)")
-        axis.minorticks_on()
-        axis.legend(loc="center left", bbox_to_anchor=(1.02, 0.5))
-        axis.set_title("CS Hoop Stress at BOP")
-        axis.grid(True, alpha=0.3)
-
-    def plot_cs_radial_stress_profile(
-        self,
-        axis: plt.Axes,
-        mfile: MFile,
-        scan: int,
-        j_cs: float,
-        b_cs_inner: float,
-    ):
-        r_cs_inner = mfile.get("r_cs_inner", scan=scan)
-        r_cs_outer = mfile.get("r_cs_outer", scan=scan)
-
-        radii = np.linspace(r_cs_inner, r_cs_outer, num=10)
-        stress_values = np.array([
-            self.calculate_cs_radial_stress(
-                r_stress_point=radius,
-                r_cs_inner=r_cs_inner,
-                r_cs_outer=r_cs_outer,
-                j_cs=j_cs,
-                b_cs_inner=b_cs_inner,
-                f_poisson_cs_structure=tfv.poisson_steel,
-            )
-            for radius in radii
-        ])
-
-        axis.plot(
-            radii,
-            stress_values / 1e6,
-            linewidth=2,
-            label="$\\sigma_{r}$,Radial Stress",
-        )
-        max_idx = np.argmax(np.abs(stress_values))
-        max_radius = radii[max_idx]
-        max_stress = stress_values[max_idx] / 1e6
-        axis.axvline(max_radius, color="black", linestyle="--", linewidth=1.0, alpha=0.7)
-        axis.axhline(max_stress, color="black", linestyle="--", linewidth=1.0, alpha=0.7)
-        axis.set_xlabel("Radial Position (m)")
-        axis.set_ylabel("Radial Stress (MPa)")
-        axis.minorticks_on()
-        axis.grid(True, alpha=0.3)
-        axis.set_title("CS Radial Stress at BOP")
-        axis.legend(loc="center left", bbox_to_anchor=(1.02, 0.5))
-
-    def plot_cs_radial_stress_contour_profile(
-        self,
-        axis: plt.Axes,
-        mfile: MFile,
-        scan: int,
-        j_cs: float,
-        b_cs_inner: float,
-        colorbar_axis: plt.Axes | None = None,
-    ):
-        r_cs_inner = mfile.get("r_cs_inner", scan=scan)
-        r_cs_outer = mfile.get("r_cs_outer", scan=scan)
-        dz_cs_full = mfile.get("dz_cs_full", scan=scan)
-
-        # Create 2D grid for contour plot: radial and vertical dimensions
-        n_radial = 50
-        radial_grid = np.linspace(r_cs_inner, r_cs_outer, n_radial)
-        height_grid = np.linspace(-dz_cs_full / 2, dz_cs_full / 2, 20)
-
-        # Create meshgrid for filled contour
-        r, z = np.meshgrid(radial_grid, height_grid)
-
-        # Calculate radial stress across the 2D grid
-        stress_data = np.zeros((len(height_grid), n_radial))
-        for i in range(len(height_grid)):
-            for j in range(n_radial):
-                stress_data[i, j] = (
-                    self.calculate_cs_radial_stress(
-                        r_stress_point=radial_grid[j],
-                        r_cs_inner=r_cs_inner,
-                        r_cs_outer=r_cs_outer,
-                        j_cs=j_cs,
-                        b_cs_inner=b_cs_inner,
-                        f_poisson_cs_structure=tfv.poisson_steel,
-                    )
-                    / 1e6
-                )
-
-        # Plot filled contour of stress distribution
-        contour_fill = axis.contourf(r, z, stress_data, levels=15, cmap="RdYlBu_r")
-        contour_lines = axis.contour(
-            r,
-            z,
-            stress_data,
-            levels=[stress_data.max()],
-            colors="black",
-            linewidths=0.5,
-            alpha=0.4,
-        )
-        axis.clabel(contour_lines, inline=True, fontsize=8)
-
-        # Plot CS outline
-        axis.plot(
-            [r_cs_inner, r_cs_inner],
-            [-dz_cs_full / 2, dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-            label="CS Inner",
-        )
-        axis.plot(
-            [r_cs_outer, r_cs_outer],
-            [-dz_cs_full / 2, dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-            label="CS Outer",
-        )
-        axis.plot(
-            [r_cs_inner, r_cs_outer], [dz_cs_full / 2, dz_cs_full / 2], "k-", linewidth=2
-        )
-        axis.plot(
-            [r_cs_inner, r_cs_outer],
-            [-dz_cs_full / 2, -dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-        )
-
-        # Use a dedicated colorbar axes when provided so the main axes width is unchanged.
-        if colorbar_axis is None:
-            cbar = axis.figure.colorbar(contour_fill, ax=axis, pad=0.02)
-        else:
-            cbar = axis.figure.colorbar(contour_fill, cax=colorbar_axis)
-        cbar.set_label("Radial Stress (MPa)")
-
-        axis.set_xlabel("R [m]")
-        axis.set_ylabel("Z [m]")
-        axis.minorticks_on()
-        axis.set_xlim(r_cs_inner * 0.9, r_cs_outer * 1.1)
-        axis.set_ylim((-dz_cs_full / 2) * 1.1, (dz_cs_full / 2) * 1.1)
-        axis.grid(True, alpha=0.3)
-
-    def plot_cs_hoop_stress_contour_profile(
-        self,
-        axis: plt.Axes,
-        mfile: MFile,
-        scan: int,
-        j_cs: float,
-        b_cs_inner: float,
-        colorbar_axis: plt.Axes | None = None,
-    ):
-        r_cs_inner = mfile.get("r_cs_inner", scan=scan)
-        r_cs_outer = mfile.get("r_cs_outer", scan=scan)
-        dz_cs_full = mfile.get("dz_cs_full", scan=scan)
-        f_a_cs_turn_steel = mfile.get("f_a_cs_turn_steel", scan=scan)
-
-        # Create 2D grid for contour plot: radial and vertical dimensions
-        n_radial = 50
-        radial_grid = np.linspace(r_cs_inner, r_cs_outer, n_radial)
-        height_grid = np.linspace(-dz_cs_full / 2, dz_cs_full / 2, 20)
-
-        # Create meshgrid for filled contour
-        r, z = np.meshgrid(radial_grid, height_grid)
-
-        # Calculate hoop stress across the 2D grid
-        stress_data = np.zeros((len(height_grid), n_radial))
-        for i in range(len(height_grid)):
-            for j in range(n_radial):
-                stress_data[i, j] = (
-                    self.calculate_cs_hoop_stress(
-                        r_stress_point=radial_grid[j],
-                        r_cs_inner=r_cs_inner,
-                        r_cs_outer=r_cs_outer,
-                        j_cs=j_cs,
-                        b_cs_inner=b_cs_inner,
-                        f_poisson_cs_structure=tfv.poisson_steel,
-                        f_a_cs_turn_steel=f_a_cs_turn_steel,
-                    )
-                    / 1e6
-                )
-
-        # Plot filled contour of stress distribution
-        contour_fill = axis.contourf(r, z, stress_data, levels=15, cmap="RdYlBu_r")
-        contour_lines = axis.contour(
-            r,
-            z,
-            stress_data,
-            levels=[stress_data.max()],
-            colors="black",
-            linewidths=0.5,
-            alpha=0.4,
-        )
-        axis.clabel(contour_lines, inline=True, fontsize=8)
-
-        # Plot CS outline
-        axis.plot(
-            [r_cs_inner, r_cs_inner],
-            [-dz_cs_full / 2, dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-            label="CS Inner",
-        )
-        axis.plot(
-            [r_cs_outer, r_cs_outer],
-            [-dz_cs_full / 2, dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-            label="CS Outer",
-        )
-        axis.plot(
-            [r_cs_inner, r_cs_outer], [dz_cs_full / 2, dz_cs_full / 2], "k-", linewidth=2
-        )
-        axis.plot(
-            [r_cs_inner, r_cs_outer],
-            [-dz_cs_full / 2, -dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-        )
-
-        # Use a dedicated colorbar axes when provided so the main axes width is unchanged.
-        if colorbar_axis is None:
-            cbar = axis.figure.colorbar(contour_fill, ax=axis, pad=0.02)
-        else:
-            cbar = axis.figure.colorbar(contour_fill, cax=colorbar_axis)
-        cbar.set_label("Hoop Stress (MPa)")
-
-        axis.set_xlabel("R [m]")
-        axis.set_ylabel("Z [m]")
-        axis.minorticks_on()
-        axis.set_xlim(r_cs_inner * 0.9, r_cs_outer * 1.1)
-        axis.set_ylim((-dz_cs_full / 2) * 1.1, (dz_cs_full / 2) * 1.1)
-        axis.grid(True, alpha=0.3)
-
-    def plot_stress_yield_locus(
-        self,
-        axis: plt.Axes,
-        mfile: MFile,
-        scan: int,
-        stress_yield: float,
-    ):
-        stress_z_cs_self_peak_midplane = mfile.get(
-            "stress_z_cs_self_peak_midplane", scan=scan
-        )
-        stress_hoop_cs_inner = mfile.get("stress_hoop_cs_inner", scan=scan)
-        stress_yield_mpa = stress_yield / 1e6
-        stress_mises_cs_peak = mfile.get("stress_mises_cs_peak", scan=scan) / 1e6
-        stress_shear_cs_peak = mfile.get("stress_shear_cs_peak", scan=scan) / 1e6
-        s1_boundary = [
-            stress_yield_mpa,
-            stress_yield_mpa,
-            0,
-            -stress_yield_mpa,
-            -stress_yield_mpa,
-            0,
-            stress_yield_mpa,
-        ]
-        s2_boundary = [
-            0,
-            stress_yield_mpa,
-            stress_yield_mpa,
-            0,
-            -stress_yield_mpa,
-            -stress_yield_mpa,
-            0,
-        ]
-
-        angles = np.linspace(0, 2 * np.pi, 300)
-        # Coordinate transformation to rotate the standard ellipse into the principal stress plane
-        vm_s1 = stress_yield_mpa * np.cos(angles)
-        vm_s2 = stress_yield_mpa * (np.cos(angles) / 2 + np.sin(angles) * np.sqrt(3) / 2)
-        axis.plot(
-            vm_s1,
-            vm_s2,
-            "b-",
-            linewidth=2,
-            label="von Mises Yield Locus\n(Ellipse)",
-        )
-
-        axis.plot(
-            s1_boundary,
-            s2_boundary,
-            color="red",
-            label="Tresca Limit Envelope",
-        )
-        axis.plot(
-            stress_hoop_cs_inner / 1e6,
-            stress_z_cs_self_peak_midplane / 1e6,
-            marker="o",
-            markersize=12,
-            color="blue",
-            label=(
-                f"CS Inboard Midplane Stress Point\n"
-                f"Tresca = {stress_shear_cs_peak:.1f} MPa\n"
-                f"von Mises = {stress_mises_cs_peak:.1f} MPa"
-            ),
-        )
-        axis.axhline(0, color="black", linewidth=1)
-        axis.axvline(0, color="black", linewidth=1)
-        axis.axhline(
-            stress_z_cs_self_peak_midplane / 1e6,
-            color="blue",
-            linewidth=1,
-            linestyle="--",
-            alpha=0.6,
-        )
-        axis.axvline(
-            stress_hoop_cs_inner / 1e6,
-            color="blue",
-            linewidth=1,
-            linestyle="--",
-            alpha=0.6,
-        )
-
-        axis.legend(loc="center left", bbox_to_anchor=(1.02, 0.5))
-        axis.grid(True, alpha=0.3)
-        axis.set_xlabel("Hoop Stress (MPa)")
-        axis.set_ylabel("Vertical Stress (MPa)")
-        axis.set_title("Stress Yield Locus")
-        axis.minorticks_on()
-
-    def plot_cs_vertical_stress_profile(
-        self,
-        axis: plt.Axes,
-        mfile: MFile,
-        scan: int,
-    ):
-        dz_cs_full = mfile.get("dz_cs_full", scan=scan)
-
-        stress_z_profile = np.array([
-            float(mfile.data[f"stress_z_cs_self_profile[{i}]"].get_scan(scan)) / 1e6
-            for i in range(20)
-        ])
-        z_positions = np.linspace(-dz_cs_full / 2, dz_cs_full / 2, len(stress_z_profile))
-
-        axis.plot(
-            stress_z_profile,
-            z_positions,
-            linewidth=2,
-            label="$\\sigma_{z}$,Vertical Stress",
-        )
-        max_idx = np.argmax(np.abs(stress_z_profile))
-        max_stress = stress_z_profile[max_idx]
-        max_z = z_positions[max_idx]
-        axis.axvline(max_stress, color="black", linestyle="--", linewidth=1.0, alpha=0.7)
-        axis.axhline(max_z, color="black", linestyle="--", linewidth=1.0, alpha=0.7)
-        axis.set_xlabel("Vertical Stress (MPa)")
-        axis.set_ylabel("Z [m]")
-        axis.minorticks_on()
-        axis.grid(True, alpha=0.3)
-        axis.set_title("CS Vertical Stress at BOP")
-        axis.legend(loc="center left", bbox_to_anchor=(1.02, 0.5))
-
-    def plot_vertical_stress_contour_profile(
-        self,
-        axis: plt.Axes,
-        mfile: MFile,
-        scan: int,
-        colorbar_axis: plt.Axes | None = None,
-    ):
-        dz_cs_full = mfile.get("dz_cs_full", scan=scan)
-        r_cs_inner = mfile.get("r_cs_inner", scan=scan)
-        r_cs_outer = mfile.get("r_cs_outer", scan=scan)
-
-        stress_z_profile = [
-            float(mfile.data[f"stress_z_cs_self_profile[{i}]"].get_scan(scan)) / 1e6
-            for i in range(20)
-        ]
-
-        # Create 2D grid for contour plot: radial and vertical dimensions
-        n_radial = 50
-        radial_grid = np.linspace(r_cs_inner, r_cs_outer, n_radial)
-        height_grid = np.linspace(-dz_cs_full / 2, dz_cs_full / 2, len(stress_z_profile))
-
-        # Create meshgrid for filled contour
-        r, z = np.meshgrid(radial_grid, height_grid)
-
-        # Interpolate stress values across radial direction (assume linear variation)
-        stress_data = np.zeros((len(stress_z_profile), n_radial))
-        for i, stress_val in enumerate(stress_z_profile):
-            stress_data[i, :] = stress_val
-
-        # Plot filled contour of stress distribution
-        contour_fill = axis.contourf(r, z, stress_data, levels=15, cmap="RdYlBu_r")
-        contour_lines = axis.contour(
-            r,
-            z,
-            stress_data,
-            levels=[stress_data.max()],
-            colors="black",
-            linewidths=0.5,
-            alpha=0.4,
-        )
-        axis.clabel(contour_lines, inline=True, fontsize=8)
-
-        # Plot CS outline
-        axis.plot(
-            [r_cs_inner, r_cs_inner],
-            [-dz_cs_full / 2, dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-            label="CS Inner",
-        )
-        axis.plot(
-            [r_cs_outer, r_cs_outer],
-            [-dz_cs_full / 2, dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-            label="CS Outer",
-        )
-        axis.plot(
-            [r_cs_inner, r_cs_outer], [dz_cs_full / 2, dz_cs_full / 2], "k-", linewidth=2
-        )
-        axis.plot(
-            [r_cs_inner, r_cs_outer],
-            [-dz_cs_full / 2, -dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-        )
-
-        # Use a dedicated colorbar axes when provided so the main axes width is unchanged.
-        if colorbar_axis is None:
-            cbar = axis.figure.colorbar(contour_fill, ax=axis, pad=0.02)
-        else:
-            cbar = axis.figure.colorbar(contour_fill, cax=colorbar_axis)
-        cbar.set_label("Vertical Stress (MPa)")
-
-        axis.set_xlabel("R [m]")
-        axis.set_ylabel("Z [m]")
-        axis.minorticks_on()
-        axis.set_xlim(r_cs_inner * 0.9, r_cs_outer * 1.1)
-        axis.set_ylim((-dz_cs_full / 2) * 1.1, (dz_cs_full / 2) * 1.1)
-        axis.grid(True, alpha=0.3)
-
-    def plot_cs_tresca_2d_contour(
-        self,
-        axis: plt.Axes,
-        mfile: MFile,
-        scan: int,
-        colorbar_axis: plt.Axes | None = None,
-    ):
-        dz_cs_full = mfile.get("dz_cs_full", scan=scan)
-        r_cs_inner = mfile.get("r_cs_inner", scan=scan)
-        r_cs_outer = mfile.get("r_cs_outer", scan=scan)
-        j_cs = mfile.get("j_cs_pulse_start", scan=scan)
-        b_cs_inner = mfile.get("b_cs_peak_pulse_start", scan=scan)
-        f_a_cs_turn_steel = mfile.get("f_a_cs_turn_steel", scan=scan)
-
-        stress_z_profile = np.array([
-            float(mfile.data[f"stress_z_cs_self_profile[{i}]"].get_scan(scan))
-            for i in range(20)
-        ])
-
-        # Create 2D grid for contour plot: radial and vertical dimensions
-        n_radial = 50
-        radial_grid = np.linspace(r_cs_inner, r_cs_outer, n_radial)
-        height_grid = np.linspace(-dz_cs_full / 2, dz_cs_full / 2, len(stress_z_profile))
-
-        # Create meshgrid for filled contour
-        r, z = np.meshgrid(radial_grid, height_grid)
-
-        # Calculate Tresca stress across the coil cross-section.
-        tresca_data = np.zeros((len(height_grid), n_radial))
-        for i, stress_z in enumerate(stress_z_profile):
-            for j, radius in enumerate(radial_grid):
-                stress_hoop = self.calculate_cs_hoop_stress(
-                    r_stress_point=radius,
-                    r_cs_inner=r_cs_inner,
-                    r_cs_outer=r_cs_outer,
-                    j_cs=j_cs,
-                    b_cs_inner=b_cs_inner,
-                    f_poisson_cs_structure=tfv.poisson_steel,
-                    f_a_cs_turn_steel=f_a_cs_turn_steel,
-                )
-                stress_radial = self.calculate_cs_radial_stress(
-                    r_stress_point=radius,
-                    r_cs_inner=r_cs_inner,
-                    r_cs_outer=r_cs_outer,
-                    j_cs=j_cs,
-                    b_cs_inner=b_cs_inner,
-                    f_poisson_cs_structure=tfv.poisson_steel,
-                )
-                tresca_data[i, j] = (
-                    calculate_tresca_stress(
-                        stress_x=stress_hoop,
-                        stress_y=stress_z,
-                        stress_z=stress_radial,
-                    )
-                    / 1e6
-                )
-
-        # Plot filled contour of Tresca stress distribution
-        contour_fill = axis.contourf(r, z, tresca_data, levels=15, cmap="RdYlBu_r")
-        contour_lines = axis.contour(
-            r,
-            z,
-            tresca_data,
-            levels=[tresca_data.max()],
-            colors="black",
-            linewidths=0.5,
-            alpha=0.4,
-        )
-        axis.clabel(contour_lines, inline=True, fontsize=8)
-
-        # Plot CS outline
-        axis.plot(
-            [r_cs_inner, r_cs_inner],
-            [-dz_cs_full / 2, dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-            label="CS Inner",
-        )
-        axis.plot(
-            [r_cs_outer, r_cs_outer],
-            [-dz_cs_full / 2, dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-            label="CS Outer",
-        )
-        axis.plot(
-            [r_cs_inner, r_cs_outer], [dz_cs_full / 2, dz_cs_full / 2], "k-", linewidth=2
-        )
-        axis.plot(
-            [r_cs_inner, r_cs_outer],
-            [-dz_cs_full / 2, -dz_cs_full / 2],
-            "k-",
-            linewidth=2,
-        )
-
-        # Use a dedicated colorbar axes when provided so the main axes width is unchanged.
-        if colorbar_axis is None:
-            cbar = axis.figure.colorbar(contour_fill, ax=axis, pad=0.02)
-        else:
-            cbar = axis.figure.colorbar(contour_fill, cax=colorbar_axis)
-        cbar.set_label("Tresca Stress (MPa)")
-
-        axis.set_xlabel("R [m]")
-        axis.set_ylabel("Z [m]")
-        axis.minorticks_on()
-        axis.set_xlim(r_cs_inner * 0.9, r_cs_outer * 1.1)
-        axis.set_ylim((-dz_cs_full / 2) * 1.1, (dz_cs_full / 2) * 1.1)
-        axis.grid(True, alpha=0.3)
-        axis.set_title("CS Tresca Stress Contour at BOP")
 
 
 def peak_b_field_at_pf_coil(
